Create token.py
In this code, we first import the hashlib module to use the SHA-256 hashing algorithm. Then, we define a function called generate_token that takes in two arguments - college_name and unique_id. Inside the function, we concatenate the two arguments to form a string called data. We then use the hashlib.sha256 function to hash the data string using the SHA-256 algorithm. We convert the hashed data to a hexadecimal string using the .hexdigest() method. Finally, we take the first 4 characters of the hexadecimal string to form the token and return it from the function. We then prompt the user to input the college name and unique ID and call the generate_token function to generate the token. The generated token is printed to the console.